requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
Django==5.1.1
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
multidict==6.1.0
numpy==2.1.1
openai==1.50.2
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
twilio==9.3.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.13.1
django-cors-headers